+++ tradplus_flutter_sdk/tradplus_sdk/CHANGELOG.md
@@ -1,20 +1,59 @@
-## 1.0.1
-1.新增地域查询接口：可用来查询是否在 欧洲 美国加州 中国
-2.新增 设置数据库大小 仅支持android
-3.开屏增加跳过，点睛开始，点睛结束 回调 仅支持iOS
-TradPlus sdk 适配版本：Android - v8.4.0.1; iOS - v8.1.0
-## 1.0.2
-Android端修复已知问题
-TradPlus sdk 适配版本：Android - v8.4.0.1; iOS - v8.1.0
-## 1.0.3
-各广告类型增加透传数据功能
-TradPlus sdk 适配版本：Android - v8.6.0.1; iOS - v8.3.20
-## 1.0.4
-1.开屏支持原生混用。开发者可以在TradPlust后台进行配置
-2.横幅支持支持原生混用。开发者可以在TradPlust后台进行配置
-3.增加了全局展示回调接口
-4.增加API支持LGPD巴西隐私政策设置接口
-TradPlus sdk 适配版本：Android - v8.8.0.1; iOS - v8.5.0
+## 1.0.9
+
+1.适配 iOS v 9.3.0
+2.适配 Android V9.6.10.1
+3.增加测试工具调用API
+
+TradPlus sdk 适配版本：Android - v9.6.10.1; iOS - v9.3.0
+
+## 1.0.8
+
+1.修复iOS已经问题。
+
+TradPlus sdk 适配版本：Android - v9.1.0.1; iOS - v8.7.0
+
+## 1.0.7
+
+1.适配Android V9.1.0.1，有API变动。
+
+TradPlus sdk 适配版本：Android - v9.1.0.1; iOS - v8.7.0
+
 ## 1.0.5
+
 1.增加广告位isloading状态的回调接口
+
 TradPlus sdk 适配版本：Android - v9.0.0.1; iOS - v8.7.0
+
+## 1.0.4
+
+1.开屏支持原生混用。开发者可以在TradPlust后台进行配置
+
+2.横幅支持支持原生混用。开发者可以在TradPlust后台进行配置
+
+3.增加了全局展示回调接口
+
+4.增加API支持LGPD巴西隐私政策设置接口
+
+TradPlus sdk 适配版本：Android - v8.8.0.1; iOS - v8.5.0
+
+## 1.0.3
+
+各广告类型增加透传数据功能
+
+TradPlus sdk 适配版本：Android - v8.6.0.1; iOS - v8.3.20
+
+## 1.0.2
+
+Android端修复已知问题
+
+TradPlus sdk 适配版本：Android - v8.4.0.1; iOS - v8.1.0
+
+## 1.0.1
+
+1.新增地域查询接口：可用来查询是否在 欧洲 美国加州 中国
+
+2.新增 设置数据库大小 仅支持android
+
+3.开屏增加跳过，点睛开始，点睛结束 回调 仅支持iOS
+
+TradPlus sdk 适配版本：Android - v8.4.0.1; iOS - v8.1.0